Moved constraint creation
